Add mining, debt burning, and sUSD swapping in Ethereum end-to-end testing
New functionalities are implemented to enrich the capabilities of Ethereum end-to-end testing. These include the addition of block mining, debt burning, and swapping to sUSD. The changes also extend the tasks in the Liquidity Provider e2e test, including elements such as account creation and sUSD deposit, thereby boosting the robustness and coverage of the testing script.