revert pyth feed back to tbtc
Unfortunately given cannon, updating this feed id leads to creating a new settle strategy.
So the solution is to push the update without any change to the pyth btc feed id. Then make the change on the chain and then update the repo to use the real btc feed id.