310 mainnets (#574)
* release 3.10 testnets

bump the versions for release

for the changes in the `core.toml` files, there seems to be a case where
the `oracle-manager` package can be deployed after the `synthetix`
system package, which can lead to unexpected failure behavior. this is
fixed by adding a `depends`.

* 3.10 mainnets

bump the versions.

also see fix in testnets pr.

* bumping all versions

* Revert "bumping all versions"

This reverts commit 70abb90557f1f648323bc9fc8adface6b015d19f.

* fix dependency in wrong place

* fix: bug in cannon prevents using latest version of om package spec

* test staleness fix

* bump versions

* correct version for base mainnet op package

* also update the pyth wrapper

---------

Co-authored-by: kaleb <kalebkeny@gmail.com>